Move NoopCostFunction to separate file
